+++ src/simtools/data_model/schema.py
@@ -13,8 +13,6 @@ from simtools.constants import (
     MODEL_PARAMETER_METASCHEMA,
     MODEL_PARAMETER_SCHEMA_PATH,
     SCHEMA_PATH,
-    SIM_TELARRAY_META_PARAMETER_METASCHEMA,
-    SIM_TELARRAY_META_PARAMETER_REGISTRY,
 )
 from simtools.data_model import format_checkers
 from simtools.dependencies import get_software_version
@@ -87,171 +85,6 @@ def get_model_parameter_schema(parameter, schema_version=None):
     """
     schema_file = get_model_parameter_schema_file(parameter)
     return load_schema(schema_file, schema_version=schema_version or "latest")
-
-
-def get_sim_telarray_meta_parameter_registry(schema_version=None, validate=True):
-    """
-    Return the sim_telarray meta-parameter registry.
-
-    Parameters
-    ----------
-    schema_version: str, optional
-        Registry schema version. If not provided, the latest version is used.
-    validate: bool
-        Validate the registry against its metaschema before returning it.
-
-    Returns
-    -------
-    dict
-        sim_telarray meta-parameter registry.
-    """
-    registry_source = load_schema(
-        SIM_TELARRAY_META_PARAMETER_REGISTRY, schema_version=schema_version or "latest"
-    )
-    if validate:
-        validate_dict_using_schema(
-            registry_source,
-            schema_file=SIM_TELARRAY_META_PARAMETER_METASCHEMA,
-            offline=True,
-            ignore_software_version=True,
-        )
-    return _build_sim_telarray_meta_parameter_registry(registry_source)
-
-
-def _build_sim_telarray_meta_parameter_registry(registry_source):
-    """Build expanded sim_telarray meta_parameter registry from source overlays."""
-    registry = {
-        key: value for key, value in registry_source.items() if key != "generated_meta_parameters"
-    }
-    meta_parameters = {}
-
-    for emitted_name, definition in registry_source.get("generated_meta_parameters", {}).items():
-        meta_parameters[emitted_name] = {"name": emitted_name, **definition}
-
-    for source_name, model_schema in names.model_parameters().items():
-        try:
-            emitted_name, _ = _get_sim_telarray_emitted_name_and_mode(model_schema)
-        except KeyError:
-            continue
-        if emitted_name in meta_parameters:
-            continue
-        definition = _build_model_parameter_meta_parameter_definition(source_name)
-        meta_parameters[definition["name"]] = definition
-
-    registry["meta_parameters"] = meta_parameters
-    return registry
-
-
-def _build_model_parameter_meta_parameter_definition(source_name, emitted_name=None):
-    """Build one sim_telarray meta_parameter definition from a model parameter schema."""
-    model_schema = get_model_parameter_schema(source_name)
-    derived_name, mode = _get_sim_telarray_emitted_name_and_mode(model_schema)
-    emitted_name = emitted_name or derived_name
-    value_schema = _get_sim_telarray_output_value_schema(
-        model_schema
-    ) or _derive_sim_telarray_value_schema(model_schema)
-    value_required = mode == "set"
-
-    return {
-        "name": emitted_name,
-        "scope": _derive_sim_telarray_scope(model_schema),
-        "mode": mode,
-        "source_type": "model_parameter",
-        "description": model_schema.get("short_description") or model_schema.get("description", ""),
-        "unit": _derive_sim_telarray_unit(model_schema),
-        "value_schema": value_schema,
-        "validation": {
-            "source_must_exist": True,
-            "mapping_must_match": True,
-            "config_value_required": value_required,
-            "emitted_value_must_match": value_required,
-        },
-    }
-
-
-def _get_sim_telarray_emitted_name_and_mode(model_schema):
-    """Return emitted sim_telarray metadata name and metadata mode."""
-    software = _get_sim_telarray_software_definition(model_schema)
-    if software.get("set_meta_parameter", False):
-        return software.get("internal_parameter_name", model_schema["name"]), "set"
-    return software.get("internal_parameter_name", model_schema["name"]), "add"
-
-
-def _get_sim_telarray_software_definition(model_schema):
-    """Return the sim_telarray simulation_software definition for a model schema."""
-    for software in model_schema.get("simulation_software", []):
-        if software.get("name") == "sim_telarray":
-            return software
-    raise KeyError(f"Model parameter without sim_telarray mapping: {model_schema['name']}")
-
-
-def _get_sim_telarray_output_value_schema(model_schema):
-    """Return an explicit sim_telarray emitted-value schema when declared."""
-    return _get_sim_telarray_software_definition(model_schema).get("output_value_schema")
-
-
-def _derive_sim_telarray_scope(model_schema):
-    """Infer sim_telarray metadata scope from instrument class."""
-    if model_schema.get("instrument", {}).get("class") == "Site":
-        return "global"
-    return "telescope"
-
-
-def _derive_sim_telarray_unit(model_schema):
-    """Infer a compact unit representation from model parameter schema data."""
-    data = model_schema.get("data", [])
-    if isinstance(data, dict):
-        return value_conversion.normalize_dimensionless_unit(data.get("unit"))
-    if not isinstance(data, list) or len(data) == 0:
-        return None
-    units = [value_conversion.normalize_dimensionless_unit(entry.get("unit")) for entry in data]
-    if len(set(units)) == 1:
-        return units[0]
-    return None
-
-
-def _derive_scalar_data_type(data_type):
-    """Return the sim_telarray scalar data_type string for a model parameter type."""
-    if _is_integer_type(data_type):
-        return "integer"
-    if _is_numeric_type(data_type):
-        return "number"
-    if data_type == "boolean":
-        return "boolean"
-    return "string"
-
-
-def _derive_sim_telarray_value_schema(model_schema):
-    """Infer value schema from model parameter schema structure."""
-    data = model_schema.get("data", [])
-    if isinstance(data, dict):
-        data = [data]
-    if not isinstance(data, list) or len(data) == 0:
-        return {"kind": "scalar", "data_type": "string"}
-
-    if len(data) > 1 and all(_is_numeric_type(entry.get("type")) for entry in data):
-        return {"kind": "fixed_numeric_tuple", "item_type": "number", "length": len(data)}
-
-    datum = data[0]
-    data_type = datum.get("type")
-
-    if data_type == "file":
-        value_schema = {"kind": "file_name"}
-        if datum.get("default") is None:
-            value_schema["allow_none_literal"] = True
-        return value_schema
-
-    return {"kind": "scalar", "data_type": _derive_scalar_data_type(data_type)}
-
-
-def _is_integer_type(data_type):
-    """Check whether a schema data type is integer-like."""
-    return data_type in {"int", "int64", "uint", "uint32", "uint64"}
-
-
-def _is_numeric_type(data_type):
-    """Check whether a schema data type is numeric-like."""
-    return _is_integer_type(data_type) or data_type in {"double", "float64", "float32"}
 
 
 def get_model_parameter_schema_version(schema_version=None):
